+++ include/DataLoader.hpp
@@ -11,15 +11,15 @@ class DataLoader
   public:
     struct NormalizationStats
     {
-        std::vector<double> min_vals;
-        std::vector<double> max_vals;
+        std::vector<float> min_vals;
+        std::vector<float> max_vals;
 
-        double eps = 1e-9;
+        float eps = 1e-9f;
     };
 
     static DataSet loadCSV(const std::string& path, int input_cols, int target_cols, char sep = ',');
     static void normalize(DataSet& data);
-    static SplitData split(const DataSet& data, double val_ratio, unsigned seed);
+    static SplitData split(const DataSet& data, float val_ratio, unsigned seed);
     static NormalizationStats calculate_stats(const DataSet& data);
     static void applyNormalization(DataSet& data, const NormalizationStats& stats);
 };
