+++ include/util.hpp
@@ -7,6 +7,6 @@
 
 Matrix matmul(const Matrix& A, const Matrix& B);
 Matrix transpose(const Matrix& A);
-Matrix outer(const std::vector<double>& a, const std::vector<double>& b);
+Matrix outer(const std::vector<float>& a, const std::vector<float>& b);
 
 #endif
